refactor(emacs): Update org-roam and gptel configurations
Updated org-roam daily node creation and finding functions to use "Roam Dailies" instead of "Roam Nodes". Added new leader keybindings for finding nodes in the brain and refreshing the nodes database. Changed the default gptel model to 'gpt-5' and updated the model switching function to accept a model name as an argument.

Modified-by: Cristian D. Moreno - Kyonax <kyonax.corp@gmail.com>